requirements.txt
python-dotenv
boto3
pyyaml
requests
pytest
black
numpy
pandas
sqlalchemy
jupyter
newsapi-python